split Register.js into separate desktop and mobile components
