+++ src/Pages/MobileRegister.css
@@ -64,7 +64,7 @@
     padding-bottom: 1em;
     display: flex;
     justify-content: center;
-    animation: stretch 0.5s ease-out 1s both, curl 1s ease-out 1.5s both; }
+    animation: stretch 0.5s ease-out 1s both, curl 1s ease-out 1.5s both; }
 
 @keyframes stretch {
   0% {
